How to test it with an .exe file
if you want to login admin screen then you should use username:Yetkili and password:admin1994. I did this project for my university lesson it's quite simple because of it. I can improve and add so many other feature. Words are Turkish but i assume it is not a problem for you.